Changes to main to allow POS tagging to be a separate step in initialization of SS data

+++ src/main.py
@@ -15,6 +15,7 @@ import torch.nn.functional as F
 
 from args import get_args
 from data_utils import build_ptb_dataset, build_sdp_dataset, build_ss_dataset
+import data_utils
 import train
 import eval
 from parser import BiaffineParser
@@ -37,6 +38,7 @@ CONLLU_FILES = []
 #PARANMT_FILE = 'para_100k.txt'
 PARANMT_FILE = 'para-nmt-5m-processed.txt'
 CHUNKS_DIR = os.path.join(DATA_DIR, '5m')
+POS_ONLY = True
 
 PAD_TOKEN = '<pad>' # XXX Weird to have out here
 UNK_TOKEN = '<unk>'
@@ -132,32 +134,43 @@ if __name__ == '__main__':
     STS_INPUT = os.path.join(STS_DIR, 'input')
     STS_GS = os.path.join(STS_DIR, 'gs')
 
-    train_ss = { 'sent_pairs': [], 'targets': []}
-    #train_path = os.path.join(data_ss_dir, f'ss_train_{PARANMT_FILE}.pkl')
     if not os.path.isdir(os.path.join(CHUNKS_DIR, 'pkl')):
         os.mkdir(os.path.join(CHUNKS_DIR, 'pkl'))
+    if not os.path.isdir(os.path.join(CHUNKS_DIR, 'tagged')):
+        os.mkdir(os.path.join(CHUNKS_DIR, 'tagged'))
+    if not os.path.isdir(os.path.join(STS_DIR, 'tagged')):
+        os.mkdir(os.path.join(STS_DIR, 'tagged'))
+
+    train_ss = {'sent_pairs': [], 'targets': []}
     if 'train' in init_ss:
         log.info(f'Initializing SS train data.')
         chunks = sorted(list(os.listdir(CHUNKS_DIR)))
-        #breakpoint()
-        #while chunks[0] != '50-para-nmt.txt':
-        #    chunks.pop(0)
-        #chunks_clean = []
-        #for chunk in chunks:
-        #    #if os.path.splitext(chunk)[-1] == '.txt':
-        #    if chunk == '50-para-nmt.txt':
-        #        chunks_clean.append(chunk)
-
+        chunks_clean = []
         for chunk in chunks:
-            print(f'Processing chunk {chunk}')
-            train_path = os.path.join(CHUNKS_DIR, 'pkl', f'{os.path.splitext(chunk)[0]}.pkl')
-            if os.path.exists(train_path):
-                if input('Path to data for chunk {chunk} exists. Overwrite? [y/n] ').lower() != 'y': 
-                    continue
+            if os.path.splitext(chunk)[-1] == '.txt':
+                chunks_clean.append(chunk)
 
-            train_ss = build_ss_dataset(os.path.join(CHUNKS_DIR, chunk), gs='', x2i=x2i)
-            with open(train_path, 'wb') as pkl:
-                pickle.dump(train_ss, pkl)
+        for chunk in chunks_clean:
+            print(f'Processing chunk {chunk}')
+            raw_sents_path = os.path.join(CHUNKS_DIR, 'tagged', f'{os.path.splitext(chunk)[0]}-tagged.pkl')
+
+            raw_sent_pairs = data_utils.paraphrase_to_sents(os.path.join(CHUNKS_DIR, chunk))
+
+            with open(raw_sents_path, 'wb') as pkl:
+                pickle.dump(raw_sent_pairs, pkl)
+
+            if not POS_ONLY:
+                with open(raw_sents_path, 'rb') as pkl:
+                    raw_sent_pairs = pickle.load(pkl)
+
+                train_path = os.path.join(CHUNKS_DIR, 'pkl', f'{os.path.splitext(chunk)[0]}.pkl')
+                if os.path.exists(train_path):
+                    if input('Path to data for chunk {chunk} exists. Overwrite? [y/n] ').lower() != 'y': 
+                        continue
+                train_ss = build_ss_dataset(raw_sent_pairs, gs='', x2i=x2i)
+                with open(train_path, 'wb') as pkl:
+                    pickle.dump(train_ss, pkl)
+
 
     elif args.train_mode > 0:
         log.info(f'Loading pickled SS train data.')
@@ -174,12 +187,22 @@ if __name__ == '__main__':
     dev_path = os.path.join(data_ss_dir, 'ss_dev.pkl')
     if 'dev' in init_ss:
         log.info(f'Initializing SS dev data.')
-        dev_ss = build_ss_dataset(
-            os.path.join(STS_INPUT, '2017'),
-            gs=os.path.join(STS_GS, '2017'),
-            x2i=x2i)
-        with open(dev_path, 'wb') as pkl:
-            pickle.dump(dev_ss, pkl)
+        
+        raw_sents_path = os.path.join(STS_DIR, 'tagged', '2017-tagged.pkl')
+        raw_sent_pairs = data_utils.paraphrase_to_sents(os.path.join(STS_INPUT, '2017'))
+
+        with open(raw_sents_path, 'wb') as pkl:
+            pickle.dump(raw_sent_pairs, pkl)
+
+        if not POS_ONLY:
+            with open(raw_sents_path, 'rb') as pkl:
+                raw_sent_pairs = pickle.load(pkl)
+            dev_ss = build_ss_dataset(
+                raw_sent_pairs,
+                gs=os.path.join(STS_GS, '2017'),
+                x2i=x2i)
+            with open(dev_path, 'wb') as pkl:
+                pickle.dump(dev_ss, pkl)
     elif args.train_mode > 0:
         log.info(f'Loading pickled SS dev data.')
         with open(dev_path, 'rb') as pkl:
@@ -189,20 +212,32 @@ if __name__ == '__main__':
     test_path = os.path.join(data_ss_dir, 'ss_test.pkl')
     if 'test' in init_ss:
         log.info(f'Initializing SS test data.')
-        test_ss = {}
+
         years = os.listdir(STS_INPUT)
         years.remove('2017')
         for year in years:
-            test_ss[year] = build_ss_dataset(
-                os.path.join(STS_INPUT, year),
-                gs=os.path.join(STS_GS, year),
-                x2i=x2i)
-        with open(test_path, 'wb') as pkl:
-            pickle.dump(test_ss, pkl)
+            raw_sents_path = os.path.join(STS_DIR, 'tagged', f'{year}-tagged.pkl')
+            raw_sent_pairs = data_utils.paraphrase_to_sents(os.path.join(STS_INPUT, year))
+
+            with open(raw_sents_path, 'wb') as pkl:
+                pickle.dump(raw_sent_pairs, pkl)
+
+            if not POS_ONLY:
+                with open(raw_sents_path, 'rb') as pkl:
+                    raw_sent_pairs = pickle.load(pkl)
+                test_ss[year] = build_ss_dataset(
+                    raw_sent_pairs,
+                    gs=os.path.join(STS_GS, year),
+                    x2i=x2i)
+                with open(test_path, 'wb') as pkl:
+                    pickle.dump(test_ss, pkl)
     elif args.evaluate_semantic:
        log.info(f'Loading pickled SS test data.')
        with open(test_path, 'rb') as pkl:
            test_ss = pickle.load(pkl)
+
+    print('Finished!')
+    exit()
 
     data_ss['train'] = train_ss
     data_ss['dev'] = dev_ss
